backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyJWT==2.10.1
